requirements.txt
mcp[cli]
PyPDF2
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv 